refact: dev panel moved to dev folder under components
